added frustrated Ising interactions

+++ MPO/MPO.jl
@@ -17,7 +17,7 @@ end
 function trMPO(params::Parameters, sample::Projector, mpo::MPO{T}) where {T<:Complex{<:AbstractFloat}} 
     A = mpo.A
     trMPO=Matrix{ComplexF64}(I, params.χ, params.χ)
-    for i::UInt8 in 1:params.N
+    for i::UInt16 in 1:params.N
         trMPO*=A[i,:,:,idx(sample,i)]
     end
     return tr(trMPO)::ComplexF64
@@ -27,7 +27,7 @@ end
 function L_MPO_strings!(L_set, sample::Projector, mpo::MPO{T}, params::Parameters, cache::Workspace) where {T<:Complex{<:AbstractFloat}} 
     A = mpo.A
     L_set[1] = cache.ID
-    for i::UInt8=1:params.N
+    for i::UInt16=1:params.N
         mul!(L_set[i+1], L_set[i], @view(A[i,:,:,idx(sample,i)]))
     end
     return L_set
@@ -37,7 +37,7 @@ end
 function R_MPO_strings!(R_set, sample::Projector, mpo::MPO{T}, params::Parameters, cache::Workspace) where {T<:Complex{<:AbstractFloat}} 
     A = mpo.A
     R_set[1] = cache.ID
-    for i::UInt8=params.N:-1:1
+    for i::UInt16=params.N:-1:1
         mul!(R_set[params.N+2-i], @view(A[i,:,:,idx(sample,i)]), R_set[params.N+1-i])
     end
     return R_set
@@ -46,9 +46,9 @@ end
 #Computes the tensor of derivatives of variational parameters: 
 function ∂MPO(sample::Projector, L_set::Vector{<:Matrix{T}}, R_set::Vector{<:Matrix{T}}, params::Parameters, cache::Workspace, mpo::MPO{T}) where {T<:Complex{<:AbstractFloat}} 
     cache.∂ = zeros(T, params.N, params.χ, params.χ, 4)
-    for m::UInt8 in 1:params.N
+    for m::UInt16 in 1:params.N
         mul!(cache.B,R_set[params.N+1-m],L_set[m])
-        for i::UInt8=1:params.χ, j::UInt8=1:params.χ
+        for i::UInt16=1:params.χ, j::UInt16=1:params.χ
             cache.∂[m,i,j,idx(sample,m)] += cache.B[j,i]
         end
     end
@@ -59,9 +59,9 @@ end
 
 function conj_∂MPO(sample::Projector, L_set::Vector{<:Matrix{T}}, R_set::Vector{<:Matrix{T}}, params::Parameters, cache::Workspace, mpo::MPO{T}) where {T<:Complex{<:AbstractFloat}} 
     cache.∂ = zeros(T, params.N, params.χ, params.χ, 4)
-    for m::UInt8 in 1:params.N
+    for m::UInt16 in 1:params.N
         mul!(cache.B,R_set[params.N+1-m],L_set[m])
-        for i::UInt8=1:params.χ, j::UInt8=1:params.χ
+        for i::UInt16=1:params.χ, j::UInt16=1:params.χ
             cache.∂[m,i,j,idx(sample,m)] += conj(cache.B[j,i])
         end
     end
@@ -74,7 +74,7 @@ end
 function trMPO(params::Parameters, sample::Projector, mpo::TI_MPO{T}) where {T<:Complex{<:AbstractFloat}} 
     A = mpo.A
     trMPO=Matrix{ComplexF64}(I, params.χ, params.χ)
-    for i::UInt8 in 1:params.N
+    for i::UInt16 in 1:params.N
         trMPO*=A[:,:,idx(sample,i)]
     end
     return tr(trMPO)::ComplexF64
@@ -84,7 +84,7 @@ end
 function L_MPO_strings!(L_set, sample::Projector, mpo::TI_MPO{T}, params::Parameters, cache::Workspace) where {T<:Complex{<:AbstractFloat}} 
     A = mpo.A
     L_set[1] = cache.ID
-    for i::UInt8=1:params.N
+    for i::UInt16=1:params.N
         mul!(L_set[i+1], L_set[i], @view(A[:,:,idx(sample,i)]))
     end
     return L_set
@@ -94,7 +94,7 @@ end
 function R_MPO_strings!(R_set, sample::Projector, mpo::TI_MPO{T}, params::Parameters, cache::Workspace) where {T<:Complex{<:AbstractFloat}} 
     A = mpo.A
     R_set[1] = cache.ID
-    for i::UInt8=params.N:-1:1
+    for i::UInt16=params.N:-1:1
         mul!(R_set[params.N+2-i], @view(A[:,:,idx(sample,i)]), R_set[params.N+1-i])
     end
     return R_set
@@ -112,11 +112,11 @@ function ∂MPO(sample::Projector, L_set::Vector{<:Matrix{T}}, R_set::Vector{<:M
     cache.∂ = zeros(T, params.χ, params.χ, 4)
     #display(cache.∂)
     #error()
-    for m::UInt8 in 1:params.N
+    for m::UInt16 in 1:params.N
         mul!(cache.B,R_set[params.N+1-m],L_set[m])
         #display(cache.B)
         #error()
-        for i::UInt8=1:params.χ, j::UInt8=1:params.χ
+        for i::UInt16=1:params.χ, j::UInt16=1:params.χ
             @inbounds cache.∂[i,j,idx(sample,m)] += cache.B[j,i]
         end
     end
@@ -127,9 +127,9 @@ end
 
 function conj_∂MPO(sample::Projector, L_set::Vector{<:Matrix{T}}, R_set::Vector{<:Matrix{T}}, params::Parameters, cache::Workspace, mpo::TI_MPO{T}) where {T<:Complex{<:AbstractFloat}} 
     cache.∂ = zeros(T, params.χ, params.χ, 4)
-    for m::UInt8 in 1:params.N
+    for m::UInt16 in 1:params.N
         mul!(cache.B,R_set[params.N+1-m],L_set[m])
-        for i::UInt8=1:params.χ, j::UInt8=1:params.χ
+        for i::UInt16=1:params.χ, j::UInt16=1:params.χ
             @inbounds cache.∂[i,j,idx(sample,m)] += conj( cache.B[j,i] )
         end
     end
@@ -140,9 +140,9 @@ end
 function m∂MPO(sample::Projector, L_set::Vector{<:Matrix{<:Complex{<:AbstractFloat}}}, 
     R_set::Vector{<:Matrix{<:Complex{<:AbstractFloat}}}, params::Parameters, cache::Workspace)
     ∂::Array{eltype(L_set[1]),3} = zeros(eltype(L_set[1]), params.χ, params.χ, 4)
-    for m::UInt8 in 1:params.N
+    for m::UInt16 in 1:params.N
         mul!(cache.B,R_set[params.N+1-m],L_set[m])
-        for i::UInt8=1:params.χ, j::UInt8=1:params.χ
+        for i::UInt16=1:params.χ, j::UInt16=1:params.χ
             @inbounds ∂[i,j,idx(sample,m)] += cache.B[j,i]
         end
     end
